pyproject.toml
[project]
name = "new-drf-lesson-28"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahmedik",email = "isa732827@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-restframework (>=0.0.1,<0.0.2)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
